Refactor QueryResults and Decoder
QueryResults has now parseMeasurement as the only method.
parseResults is a normal function and is deprecated now.

QueryParams now stores a Decoder so that the user can configure
a Decoder at runtime rather than hardcoding into QueryResults
instances. Due to the lack of impredicative types in GHC
Decoder needs to be monomorphic to be usable with lens.