Add Ignored and Empty, and deprecate QueryResults for Void
The QueryResults instance for Void is not very clear from its name
if it accepts non-empty results or not. This commit introduces two
types to distinguish the usecases. Empty is used to indicate that
the expected results are always empty. Ignored is used to accept
any results and to always return an empty vector.

Now the QueryResults instance for Void is deprecated.